requirements.txt
astroid==2.2.5
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
fire==0.1.3
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
idna==2.8
importlib-metadata==0.18
isort==4.3.21
jedi==0.13.3
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.2
parso==0.5.0
pluggy==0.12.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pydocstyle==4.0.0
pyflakes==2.1.1
pylint==2.3.1
python-jsonrpc-server==0.2.0
python-language-server==0.27.0
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.14.0
rsa==4.0
six==1.12.0
snowballstemmer==1.9.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
yapf==0.28.0
zipp==0.5.2
